city and countrie

+++ artics-laravel/database/migrations/2020_06_23_000001_create_cities_table.php
@@ -15,7 +15,7 @@ class CreateCitiesTable extends Migration
     {
         Schema::create('cities', function (Blueprint $table) {
             $table->primary('id');
-            $table->integer('id')->unique();
+            $table->bigIncrements('id')->unique();
             $table->integer('postalCode');//must be unique for a country
             $table->string('cityName');
             $table->string('canton',2)->nullable();
